api/requirements.txt
litestar>=2.0
uvicorn[standard]>=0.23
psycopg2-binary>=2.9
asyncpg>=0.28
ollama>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.13
pdfplumber>=0.9
pandas>=2.0
openpyxl>=3.0
beautifulsoup4>=4.10
requests>=2.28
pydantic>=2.0
python-dotenv>=0.21
aiofiles>=23.0
pytest>=8.0